Orb collection, light puzzle, main menu
Orbs can now be collected, and the follower will knock down the one at the top of the pillar in the entrance area.
The light columns have been tweaked to aim only at columns that can be next in the puzzle and never backwards or off into space.
Boat model changed on main menu.